scripts/requirements.txt
# InstaDish RAG Dependencies
sentence-transformers
faiss-cpu
numpy
pandas
